move Type and BaseType to separate headers to avoid circular dep
